backend-master/api/requirements.txt
flask
python-socketio[client]
gevent
gevent-websocket
flask_cors
sqlalchemy
flask_sqlalchemy
flask_bcrypt